modularizes detection component
