--- src/componentsSaurav/More.jsx
@@ -1,14 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
-
-const More = () => {
-  return (
-    <View>
-      <Text>I think we should add drawer here, or can do on top right as well , but we need one component here, maybe favourite places or wishlist or saved pa</Text>
-    </View>
-  )
-}
-
-export default More
-
-const styles = StyleSheet.create({})
