+++ src/componentsSaurav/customComponents/SkeletonScreen.jsx
@@ -1,13 +1,13 @@
 import { View, Text } from 'react-native'
 import React from 'react'
-import MyLoader from './MyContentLoader'
+import MyLoader from './DetectionLoaderSkeleton'
 
 const SkeletonScreen = () => {
 
   return (
     <View style={{height:'100%', backgroundColor:'white'}}>
       <MyLoader/>
-      <MyLoader/>
+      {/* <MyLoader/> */}
     </View>
   )
 }
